@@ -8,6 +8,8 @@ import (
 )
 
 func TestForceNewIf(t *testing.T) {
+	t.Parallel()
+
 	t.Run("true", func(t *testing.T) {
 		var condCalls int
 		var gotOld1, gotNew1, gotOld2, gotNew2 string
@@ -27,15 +29,15 @@ func TestForceNewIf(t *testing.T) {
 				// updated.
 
 				condCalls++
-				old, new := d.GetChange("foo")
+				oldValue, newValue := d.GetChange("foo")
 
 				switch condCalls {
 				case 1:
-					gotOld1 = old.(string)
-					gotNew1 = new.(string)
+					gotOld1 = oldValue.(string)
+					gotNew1 = newValue.(string)
 				case 2:
-					gotOld2 = old.(string)
-					gotNew2 = new.(string)
+					gotOld2 = oldValue.(string)
+					gotNew2 = newValue.(string)
 				}
 
 				return true
@@ -75,6 +77,24 @@ func TestForceNewIf(t *testing.T) {
 
 		if !diff.Attributes["foo"].RequiresNew {
 			t.Error("Attribute 'foo' is not marked as RequiresNew")
+		}
+	})
+	t.Run("true-non-existent-attribute", func(t *testing.T) {
+		provider := testProvider(
+			map[string]*schema.Schema{},
+			ForceNewIf("non-existent", func(_ context.Context, d *schema.ResourceDiff, meta interface{}) bool {
+				return true
+			}),
+		)
+
+		_, err := testDiff(
+			provider,
+			map[string]string{},
+			map[string]string{},
+		)
+
+		if err != nil {
+			t.Fatalf("unexpected error: %s", err)
 		}
 	})
 	t.Run("false", func(t *testing.T) {
@@ -90,9 +110,9 @@ func TestForceNewIf(t *testing.T) {
 			},
 			ForceNewIf("foo", func(ctx context.Context, d *schema.ResourceDiff, meta interface{}) bool {
 				condCalls++
-				old, new := d.GetChange("foo")
-				gotOld = old.(string)
-				gotNew = new.(string)
+				oldValue, newValue := d.GetChange("foo")
+				gotOld = oldValue.(string)
+				gotNew = newValue.(string)
 
 				return false
 			}),
@@ -127,9 +147,29 @@ func TestForceNewIf(t *testing.T) {
 			t.Error("Attribute 'foo' is marked as RequiresNew, but should not be")
 		}
 	})
+	t.Run("false-non-existent-attribute", func(t *testing.T) {
+		provider := testProvider(
+			map[string]*schema.Schema{},
+			ForceNewIf("non-existent", func(_ context.Context, d *schema.ResourceDiff, meta interface{}) bool {
+				return false
+			}),
+		)
+
+		_, err := testDiff(
+			provider,
+			map[string]string{},
+			map[string]string{},
+		)
+
+		if err != nil {
+			t.Fatalf("unexpected error: %s", err)
+		}
+	})
 }
 
 func TestForceNewIfChange(t *testing.T) {
+	t.Parallel()
+
 	t.Run("true", func(t *testing.T) {
 		var condCalls int
 		var gotOld1, gotNew1, gotOld2, gotNew2 string
@@ -141,7 +181,7 @@ func TestForceNewIfChange(t *testing.T) {
 					Optional: true,
 				},
 			},
-			ForceNewIfChange("foo", func(_ context.Context, old, new, meta interface{}) bool {
+			ForceNewIfChange("foo", func(_ context.Context, oldValue, newValue, meta interface{}) bool {
 				// When we set "ForceNew", our CustomizeDiff function is actually
 				// called a second time to construct the "create" portion of
 				// the replace diff. On the second call, the old value is masked
@@ -152,11 +192,11 @@ func TestForceNewIfChange(t *testing.T) {
 
 				switch condCalls {
 				case 1:
-					gotOld1 = old.(string)
-					gotNew1 = new.(string)
+					gotOld1 = oldValue.(string)
+					gotNew1 = newValue.(string)
 				case 2:
-					gotOld2 = old.(string)
-					gotNew2 = new.(string)
+					gotOld2 = oldValue.(string)
+					gotNew2 = newValue.(string)
 				}
 
 				return true
@@ -198,6 +238,24 @@ func TestForceNewIfChange(t *testing.T) {
 			t.Error("Attribute 'foo' is not marked as RequiresNew")
 		}
 	})
+	t.Run("true-non-existent-attribute", func(t *testing.T) {
+		provider := testProvider(
+			map[string]*schema.Schema{},
+			ForceNewIfChange("foo", func(_ context.Context, oldValue, newValue, meta interface{}) bool {
+				return true
+			}),
+		)
+
+		_, err := testDiff(
+			provider,
+			map[string]string{},
+			map[string]string{},
+		)
+
+		if err != nil {
+			t.Fatalf("unexpected error: %s", err)
+		}
+	})
 	t.Run("false", func(t *testing.T) {
 		var condCalls int
 		var gotOld, gotNew string
@@ -209,10 +267,10 @@ func TestForceNewIfChange(t *testing.T) {
 					Optional: true,
 				},
 			},
-			ForceNewIfChange("foo", func(_ context.Context, old, new, meta interface{}) bool {
+			ForceNewIfChange("foo", func(_ context.Context, oldValue, newValue, meta interface{}) bool {
 				condCalls++
-				gotOld = old.(string)
-				gotNew = new.(string)
+				gotOld = oldValue.(string)
+				gotNew = newValue.(string)
 
 				return false
 			}),
@@ -245,6 +303,24 @@ func TestForceNewIfChange(t *testing.T) {
 
 		if diff.Attributes["foo"].RequiresNew {
 			t.Error("Attribute 'foo' is marked as RequiresNew, but should not be")
+		}
+	})
+	t.Run("false-non-existent-attribute", func(t *testing.T) {
+		provider := testProvider(
+			map[string]*schema.Schema{},
+			ForceNewIfChange("foo", func(_ context.Context, oldValue, newValue, meta interface{}) bool {
+				return true
+			}),
+		)
+
+		_, err := testDiff(
+			provider,
+			map[string]string{},
+			map[string]string{},
+		)
+
+		if err != nil {
+			t.Fatalf("unexpected error: %s", err)
 		}
 	})
 }
