@@ -18,16 +18,16 @@ func testProvider(s map[string]*schema.Schema, cd schema.CustomizeDiffFunc) *sch
 	}
 }
 
-func testDiff(provider *schema.Provider, old, new map[string]string) (*terraform.InstanceDiff, error) {
-	newI := make(map[string]interface{}, len(new))
-	for k, v := range new {
+func testDiff(provider *schema.Provider, oldValue, newValue map[string]string) (*terraform.InstanceDiff, error) {
+	newI := make(map[string]interface{}, len(newValue))
+	for k, v := range newValue {
 		newI[k] = v
 	}
 
 	return provider.ResourcesMap["test"].Diff(
 		context.Background(),
 		&terraform.InstanceState{
-			Attributes: old,
+			Attributes: oldValue,
 		},
 		&terraform.ResourceConfig{
 			Config: newI,
