@@ -1,14 +1,21 @@
 package resource
 
 import (
+	"context"
 	"errors"
 	"fmt"
-	"log"
 
+	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/logging"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
-func testStepTaint(state *terraform.State, step TestStep) error {
+func testStepTaint(ctx context.Context, state *terraform.State, step TestStep) error {
+	if len(step.Taint) == 0 {
+		return nil
+	}
+
+	logging.HelperResourceTrace(ctx, fmt.Sprintf("Using TestStep Taint: %v", step.Taint))
+
 	for _, p := range step.Taint {
 		m := state.RootModule()
 		if m == nil {
@@ -18,7 +25,7 @@ func testStepTaint(state *terraform.State, step TestStep) error {
 		if !ok {
 			return fmt.Errorf("resource %q not found in state", p)
 		}
-		log.Printf("[WARN] Test: Explicitly tainting resource %q", p)
+		logging.HelperResourceWarn(ctx, fmt.Sprintf("Explicitly tainting resource %q", p))
 		rs.Taint()
 	}
 	return nil
