@@ -1,8 +1,8 @@
 package resource
 
 import (
+	"context"
 	"fmt"
-	"log"
 	"reflect"
 	"strings"
 
@@ -10,63 +10,77 @@ import (
 	tfjson "github.com/hashicorp/terraform-json"
 	testing "github.com/mitchellh/go-testing-interface"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/logging"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/plugintest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
-func runPostTestDestroy(t testing.T, c TestCase, wd *plugintest.WorkingDir, factories map[string]func() (*schema.Provider, error), v5factories map[string]func() (tfprotov5.ProviderServer, error), v6factories map[string]func() (tfprotov6.ProviderServer, error), statePreDestroy *terraform.State) error {
+func runPostTestDestroy(ctx context.Context, t testing.T, c TestCase, wd *plugintest.WorkingDir, providers *providerFactories, statePreDestroy *terraform.State) error {
 	t.Helper()
 
-	err := runProviderCommand(t, func() error {
-		return wd.Destroy()
-	}, wd, providerFactories{
-		legacy:  factories,
-		protov5: v5factories,
-		protov6: v6factories})
+	err := runProviderCommand(ctx, t, func() error {
+		return wd.Destroy(ctx)
+	}, wd, providers)
 	if err != nil {
 		return err
 	}
 
 	if c.CheckDestroy != nil {
+		logging.HelperResourceTrace(ctx, "Using TestCase CheckDestroy")
+		logging.HelperResourceDebug(ctx, "Calling TestCase CheckDestroy")
+
 		if err := c.CheckDestroy(statePreDestroy); err != nil {
 			return err
 		}
+
+		logging.HelperResourceDebug(ctx, "Called TestCase CheckDestroy")
 	}
 
 	return nil
 }
 
-func runNewTest(t testing.T, c TestCase, helper *plugintest.Helper) {
+func runNewTest(ctx context.Context, t testing.T, c TestCase, helper *plugintest.Helper) {
 	t.Helper()
 
 	spewConf := spew.NewDefaultConfig()
 	spewConf.SortKeys = true
-	wd := helper.RequireNewWorkingDir(t)
+	wd := helper.RequireNewWorkingDir(ctx, t)
+
+	ctx = logging.TestTerraformPathContext(ctx, wd.GetHelper().TerraformExecPath())
+	ctx = logging.TestWorkingDirectoryContext(ctx, wd.GetHelper().WorkingDirectory())
+
+	providers := &providerFactories{
+		legacy:  c.ProviderFactories,
+		protov5: c.ProtoV5ProviderFactories,
+		protov6: c.ProtoV6ProviderFactories,
+	}
 
 	defer func() {
 		var statePreDestroy *terraform.State
 		var err error
-		err = runProviderCommand(t, func() error {
-			statePreDestroy, err = getState(t, wd)
+		err = runProviderCommand(ctx, t, func() error {
+			statePreDestroy, err = getState(ctx, t, wd)
 			if err != nil {
 				return err
 			}
 			return nil
-		}, wd, providerFactories{
-			legacy:  c.ProviderFactories,
-			protov5: c.ProtoV5ProviderFactories,
-			protov6: c.ProtoV6ProviderFactories})
+		}, wd, providers)
 		if err != nil {
+			logging.HelperResourceError(ctx,
+				"Error retrieving state, there may be dangling resources",
+				map[string]interface{}{logging.KeyError: err},
+			)
 			t.Fatalf("Error retrieving state, there may be dangling resources: %s", err.Error())
 			return
 		}
 
 		if !stateIsEmpty(statePreDestroy) {
-			err := runPostTestDestroy(t, c, wd, c.ProviderFactories, c.ProtoV5ProviderFactories, c.ProtoV6ProviderFactories, statePreDestroy)
+			err := runPostTestDestroy(ctx, t, c, wd, providers, statePreDestroy)
 			if err != nil {
+				logging.HelperResourceError(ctx,
+					"Error running post-test destroy, there may be dangling resources",
+					map[string]interface{}{logging.KeyError: err},
+				)
 				t.Fatalf("Error running post-test destroy, there may be dangling resources: %s", err.Error())
 			}
 		}
@@ -74,95 +88,233 @@ func runNewTest(t testing.T, c TestCase, helper *plugintest.Helper) {
 		wd.Close()
 	}()
 
-	providerCfg, err := testProviderConfig(c)
-	if err != nil {
-		t.Fatal(err)
+	if c.hasProviders(ctx) {
+		err := wd.SetConfig(ctx, c.providerConfig(ctx))
+
+		if err != nil {
+			logging.HelperResourceError(ctx,
+				"TestCase error setting provider configuration",
+				map[string]interface{}{logging.KeyError: err},
+			)
+			t.Fatalf("TestCase error setting provider configuration: %s", err)
+		}
+
+		err = runProviderCommand(ctx, t, func() error {
+			return wd.Init(ctx)
+		}, wd, providers)
+
+		if err != nil {
+			logging.HelperResourceError(ctx,
+				"TestCase error running init",
+				map[string]interface{}{logging.KeyError: err},
+			)
+			t.Fatalf("TestCase error running init: %s", err.Error())
+		}
 	}
 
-	err = wd.SetConfig(providerCfg)
-	if err != nil {
-		t.Fatalf("Error setting test config: %s", err)
-	}
-	err = runProviderCommand(t, func() error {
-		return wd.Init()
-	}, wd, providerFactories{
-		legacy:  c.ProviderFactories,
-		protov5: c.ProtoV5ProviderFactories,
-		protov6: c.ProtoV6ProviderFactories})
-	if err != nil {
-		t.Fatalf("Error running init: %s", err.Error())
-		return
-	}
+	logging.HelperResourceDebug(ctx, "Starting TestSteps")
 
 	// use this to track last step succesfully applied
 	// acts as default for import tests
 	var appliedCfg string
 
-	for i, step := range c.Steps {
+	for stepIndex, step := range c.Steps {
+		stepNumber := stepIndex + 1 // 1-based indexing for humans
+		ctx = logging.TestStepNumberContext(ctx, stepNumber)
+
+		logging.HelperResourceDebug(ctx, "Starting TestStep")
+
 		if step.PreConfig != nil {
+			logging.HelperResourceDebug(ctx, "Calling TestStep PreConfig")
 			step.PreConfig()
+			logging.HelperResourceDebug(ctx, "Called TestStep PreConfig")
 		}
 
 		if step.SkipFunc != nil {
+			logging.HelperResourceDebug(ctx, "Calling TestStep SkipFunc")
+
 			skip, err := step.SkipFunc()
 			if err != nil {
-				t.Fatal(err)
+				logging.HelperResourceError(ctx,
+					"Error calling TestStep SkipFunc",
+					map[string]interface{}{logging.KeyError: err},
+				)
+				t.Fatalf("Error calling TestStep SkipFunc: %s", err.Error())
 			}
+
+			logging.HelperResourceDebug(ctx, "Called TestStep SkipFunc")
+
 			if skip {
-				log.Printf("[WARN] Skipping step %d/%d", i+1, len(c.Steps))
+				t.Logf("Skipping step %d/%d due to SkipFunc", stepNumber, len(c.Steps))
+				logging.HelperResourceWarn(ctx, "Skipping TestStep due to SkipFunc")
 				continue
 			}
 		}
 
+		if step.Config != "" && !step.Destroy && len(step.Taint) > 0 {
+			var state *terraform.State
+
+			err := runProviderCommand(ctx, t, func() error {
+				var err error
+
+				state, err = getState(ctx, t, wd)
+
+				if err != nil {
+					return err
+				}
+
+				return nil
+			}, wd, providers)
+
+			if err != nil {
+				logging.HelperResourceError(ctx,
+					"TestStep error reading prior state before tainting resources",
+					map[string]interface{}{logging.KeyError: err},
+				)
+				t.Fatalf("TestStep %d/%d error reading prior state before tainting resources: %s", stepNumber, len(c.Steps), err)
+			}
+
+			err = testStepTaint(ctx, state, step)
+
+			if err != nil {
+				logging.HelperResourceError(ctx,
+					"TestStep error tainting resources",
+					map[string]interface{}{logging.KeyError: err},
+				)
+				t.Fatalf("TestStep %d/%d error tainting resources: %s", stepNumber, len(c.Steps), err)
+			}
+		}
+
+		if step.hasProviders(ctx) {
+			providers = &providerFactories{
+				legacy:  sdkProviderFactories(c.ProviderFactories).merge(step.ProviderFactories),
+				protov5: protov5ProviderFactories(c.ProtoV5ProviderFactories).merge(step.ProtoV5ProviderFactories),
+				protov6: protov6ProviderFactories(c.ProtoV6ProviderFactories).merge(step.ProtoV6ProviderFactories),
+			}
+
+			providerCfg := step.providerConfig(ctx)
+
+			err := wd.SetConfig(ctx, providerCfg)
+
+			if err != nil {
+				logging.HelperResourceError(ctx,
+					"TestStep error setting provider configuration",
+					map[string]interface{}{logging.KeyError: err},
+				)
+				t.Fatalf("TestStep %d/%d error setting test provider configuration: %s", stepNumber, len(c.Steps), err)
+			}
+
+			err = runProviderCommand(
+				ctx,
+				t,
+				func() error {
+					return wd.Init(ctx)
+				},
+				wd,
+				providers,
+			)
+
+			if err != nil {
+				logging.HelperResourceError(ctx,
+					"TestStep error running init",
+					map[string]interface{}{logging.KeyError: err},
+				)
+				t.Fatalf("TestStep %d/%d running init: %s", stepNumber, len(c.Steps), err.Error())
+				return
+			}
+		}
+
 		if step.ImportState {
-			err := testStepNewImportState(t, c, helper, wd, step, appliedCfg)
+			logging.HelperResourceTrace(ctx, "TestStep is ImportState mode")
+
+			err := testStepNewImportState(ctx, t, helper, wd, step, appliedCfg, providers)
 			if step.ExpectError != nil {
+				logging.HelperResourceDebug(ctx, "Checking TestStep ExpectError")
 				if err == nil {
-					t.Fatalf("Step %d/%d error running import: expected an error but got none", i+1, len(c.Steps))
+					logging.HelperResourceError(ctx,
+						"Error running import: expected an error but got none",
+					)
+					t.Fatalf("Step %d/%d error running import: expected an error but got none", stepNumber, len(c.Steps))
 				}
 				if !step.ExpectError.MatchString(err.Error()) {
-					t.Fatalf("Step %d/%d error running import, expected an error with pattern (%s), no match on: %s", i+1, len(c.Steps), step.ExpectError.String(), err)
+					logging.HelperResourceError(ctx,
+						fmt.Sprintf("Error running import: expected an error with pattern (%s)", step.ExpectError.String()),
+						map[string]interface{}{logging.KeyError: err},
+					)
+					t.Fatalf("Step %d/%d error running import, expected an error with pattern (%s), no match on: %s", stepNumber, len(c.Steps), step.ExpectError.String(), err)
 				}
 			} else {
 				if err != nil && c.ErrorCheck != nil {
+					logging.HelperResourceDebug(ctx, "Calling TestCase ErrorCheck")
 					err = c.ErrorCheck(err)
+					logging.HelperResourceDebug(ctx, "Called TestCase ErrorCheck")
 				}
 				if err != nil {
-					t.Fatalf("Step %d/%d error running import: %s", i+1, len(c.Steps), err)
+					logging.HelperResourceError(ctx,
+						"Error running import",
+						map[string]interface{}{logging.KeyError: err},
+					)
+					t.Fatalf("Step %d/%d error running import: %s", stepNumber, len(c.Steps), err)
 				}
 			}
+
+			logging.HelperResourceDebug(ctx, "Finished TestStep")
+
 			continue
 		}
 
 		if step.Config != "" {
-			err := testStepNewConfig(t, c, wd, step)
+			logging.HelperResourceTrace(ctx, "TestStep is Config mode")
+
+			err := testStepNewConfig(ctx, t, c, wd, step, providers)
 			if step.ExpectError != nil {
+				logging.HelperResourceDebug(ctx, "Checking TestStep ExpectError")
+
 				if err == nil {
-					t.Fatalf("Step %d/%d, expected an error but got none", i+1, len(c.Steps))
+					logging.HelperResourceError(ctx,
+						"Expected an error but got none",
+					)
+					t.Fatalf("Step %d/%d, expected an error but got none", stepNumber, len(c.Steps))
 				}
 				if !step.ExpectError.MatchString(err.Error()) {
-					t.Fatalf("Step %d/%d, expected an error with pattern, no match on: %s", i+1, len(c.Steps), err)
+					logging.HelperResourceError(ctx,
+						fmt.Sprintf("Expected an error with pattern (%s)", step.ExpectError.String()),
+						map[string]interface{}{logging.KeyError: err},
+					)
+					t.Fatalf("Step %d/%d, expected an error with pattern, no match on: %s", stepNumber, len(c.Steps), err)
 				}
 			} else {
 				if err != nil && c.ErrorCheck != nil {
+					logging.HelperResourceDebug(ctx, "Calling TestCase ErrorCheck")
+
 					err = c.ErrorCheck(err)
+
+					logging.HelperResourceDebug(ctx, "Called TestCase ErrorCheck")
 				}
 				if err != nil {
-					t.Fatalf("Step %d/%d error: %s", i+1, len(c.Steps), err)
+					logging.HelperResourceError(ctx,
+						"Unexpected error",
+						map[string]interface{}{logging.KeyError: err},
+					)
+					t.Fatalf("Step %d/%d error: %s", stepNumber, len(c.Steps), err)
 				}
 			}
+
 			appliedCfg = step.Config
+
+			logging.HelperResourceDebug(ctx, "Finished TestStep")
+
 			continue
 		}
 
-		t.Fatal("Unsupported test mode")
+		t.Fatalf("Step %d/%d, unsupported test mode", stepNumber, len(c.Steps))
 	}
 }
 
-func getState(t testing.T, wd *plugintest.WorkingDir) (*terraform.State, error) {
+func getState(ctx context.Context, t testing.T, wd *plugintest.WorkingDir) (*terraform.State, error) {
 	t.Helper()
 
-	jsonState, err := wd.State()
+	jsonState, err := wd.State(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -188,7 +340,7 @@ func planIsEmpty(plan *tfjson.Plan) bool {
 	return true
 }
 
-func testIDRefresh(c TestCase, t testing.T, wd *plugintest.WorkingDir, step TestStep, r *terraform.ResourceState) error {
+func testIDRefresh(ctx context.Context, t testing.T, c TestCase, wd *plugintest.WorkingDir, step TestStep, r *terraform.ResourceState, providers *providerFactories) error {
 	t.Helper()
 
 	spewConf := spew.NewDefaultConfig()
@@ -202,36 +354,29 @@ func testIDRefresh(c TestCase, t testing.T, wd *plugintest.WorkingDir, step Test
 
 	// Temporarily set the config to a minimal provider config for the refresh
 	// test. After the refresh we can reset it.
-	cfg, err := testProviderConfig(c)
-	if err != nil {
-		return err
-	}
-	err = wd.SetConfig(cfg)
+	err := wd.SetConfig(ctx, c.providerConfig(ctx))
 	if err != nil {
 		t.Fatalf("Error setting import test config: %s", err)
 	}
 	defer func() {
-		err = wd.SetConfig(step.Config)
+		err = wd.SetConfig(ctx, step.Config)
 		if err != nil {
 			t.Fatalf("Error resetting test config: %s", err)
 		}
 	}()
 
 	// Refresh!
-	err = runProviderCommand(t, func() error {
-		err = wd.Refresh()
+	err = runProviderCommand(ctx, t, func() error {
+		err = wd.Refresh(ctx)
 		if err != nil {
 			t.Fatalf("Error running terraform refresh: %s", err)
 		}
-		state, err = getState(t, wd)
+		state, err = getState(ctx, t, wd)
 		if err != nil {
 			return err
 		}
 		return nil
-	}, wd, providerFactories{
-		legacy:  c.ProviderFactories,
-		protov5: c.ProtoV5ProviderFactories,
-		protov6: c.ProtoV6ProviderFactories})
+	}, wd, providers)
 	if err != nil {
 		return err
 	}
@@ -246,6 +391,11 @@ func testIDRefresh(c TestCase, t testing.T, wd *plugintest.WorkingDir, step Test
 	}
 	actual := actualR.Primary.Attributes
 	expected := r.Primary.Attributes
+
+	if len(c.IDRefreshIgnore) > 0 {
+		logging.HelperResourceTrace(ctx, fmt.Sprintf("Using TestCase IDRefreshIgnore: %v", c.IDRefreshIgnore))
+	}
+
 	// Remove fields we're ignoring
 	for _, v := range c.IDRefreshIgnore {
 		for k := range actual {
