@@ -14,6 +14,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/configs/configschema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/logging"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/meta"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
@@ -265,7 +266,7 @@ func (p *Provider) Configure(ctx context.Context, c *terraform.ResourceConfig) d
 	}
 
 	if p.configured {
-		log.Printf("[WARN] Previously configured provider being re-configured. This can cause issues in concurrent testing if the configurations are not equal.")
+		logging.HelperSchemaWarn(ctx, "Previously configured provider being re-configured. This can cause issues in concurrent testing if the configurations are not equal.")
 	}
 
 	sm := schemaMap(p.Schema)
@@ -378,11 +379,15 @@ func (p *Provider) ImportState(
 	results := []*ResourceData{data}
 	if r.Importer.State != nil || r.Importer.StateContext != nil {
 		var err error
+		logging.HelperSchemaTrace(ctx, "Calling downstream")
+
 		if r.Importer.StateContext != nil {
 			results, err = r.Importer.StateContext(ctx, data, p.meta)
 		} else {
 			results, err = r.Importer.State(data, p.meta)
 		}
+		logging.HelperSchemaTrace(ctx, "Called downstream")
+
 		if err != nil {
 			return nil, err
 		}
@@ -391,6 +396,21 @@ func (p *Provider) ImportState(
 	// Convert the results to InstanceState values and return it
 	states := make([]*terraform.InstanceState, len(results))
 	for i, r := range results {
+		if r == nil {
+			return nil, fmt.Errorf("The provider returned a missing resource during ImportResourceState. " +
+				"This is generally a bug in the resource implementation for import. " +
+				"Resource import code should return an error for missing resources and skip returning a missing or empty ResourceData. " +
+				"Please report this to the provider developers.")
+		}
+
+		if r.Id() == "" {
+			return nil, fmt.Errorf("The provider returned a resource missing an identifier during ImportResourceState. " +
+				"This is generally a bug in the resource implementation for import. " +
+				"Resource import code should not call d.SetId(\"\") or create an empty ResourceData. " +
+				"If the resource is missing, instead return an error. " +
+				"Please report this to the provider developers.")
+		}
+
 		states[i] = r.State()
 	}
 
@@ -398,10 +418,10 @@ func (p *Provider) ImportState(
 	// isn't obvious so we circumvent that with a friendlier error.
 	for _, s := range states {
 		if s == nil {
-			return nil, fmt.Errorf(
-				"nil entry in ImportState results. This is always a bug with\n" +
-					"the resource that is being imported. Please report this as\n" +
-					"a bug to Terraform.")
+			return nil, fmt.Errorf("The provider returned a missing resource during ImportResourceState. " +
+				"This is generally a bug in the resource implementation for import. " +
+				"Resource import code should return an error for missing resources. " +
+				"Please report this to the provider developers.")
 		}
 	}
 
