@@ -270,16 +270,16 @@ func (d *ResourceDiff) GetChangedKeysPrefix(prefix string) []string {
 // diffChange helps to implement resourceDiffer and derives its change values
 // from ResourceDiff's own change data, in addition to existing diff, config, and state.
 func (d *ResourceDiff) diffChange(key string) (interface{}, interface{}, bool, bool, bool) {
-	old, new, customized := d.getChange(key)
+	oldValue, newValue, customized := d.getChange(key)
 
-	if !old.Exists {
-		old.Value = nil
+	if !oldValue.Exists {
+		oldValue.Value = nil
 	}
-	if !new.Exists || d.removed(key) {
-		new.Value = nil
+	if !newValue.Exists || d.removed(key) {
+		newValue.Value = nil
 	}
 
-	return old.Value, new.Value, !reflect.DeepEqual(old.Value, new.Value), new.Computed, customized
+	return oldValue.Value, newValue.Value, !reflect.DeepEqual(oldValue.Value, newValue.Value), newValue.Computed, customized
 }
 
 // SetNew is used to set a new diff value for the mentioned key. The value must
@@ -308,12 +308,12 @@ func (d *ResourceDiff) SetNewComputed(key string) error {
 }
 
 // setDiff performs common diff setting behaviour.
-func (d *ResourceDiff) setDiff(key string, new interface{}, computed bool) error {
+func (d *ResourceDiff) setDiff(key string, newValue interface{}, computed bool) error {
 	if err := d.clear(key); err != nil {
 		return err
 	}
 
-	if err := d.newWriter.WriteField(strings.Split(key, "."), new, computed); err != nil {
+	if err := d.newWriter.WriteField(strings.Split(key, "."), newValue, computed); err != nil {
 		return fmt.Errorf("Cannot set new diff value for key %s: %s", key, err)
 	}
 
@@ -374,8 +374,8 @@ func (d *ResourceDiff) Get(key string) interface{} {
 // results from the exact levels for the new diff, then from state and diff as
 // per normal.
 func (d *ResourceDiff) GetChange(key string) (interface{}, interface{}) {
-	old, new, _ := d.getChange(key)
-	return old.Value, new.Value
+	oldValue, newValue, _ := d.getChange(key)
+	return oldValue.Value, newValue.Value
 }
 
 // GetOk functions the same way as ResourceData.GetOk, but it also checks the
@@ -422,19 +422,29 @@ func (d *ResourceDiff) NewValueKnown(key string) bool {
 	return !r.Computed
 }
 
+// HasChanges returns whether or not any of the given keys has been changed.
+func (d *ResourceDiff) HasChanges(keys ...string) bool {
+	for _, key := range keys {
+		if d.HasChange(key) {
+			return true
+		}
+	}
+	return false
+}
+
 // HasChange checks to see if there is a change between state and the diff, or
 // in the overridden diff.
 func (d *ResourceDiff) HasChange(key string) bool {
-	old, new := d.GetChange(key)
+	oldValue, newValue := d.GetChange(key)
 
 	// If the type implements the Equal interface, then call that
 	// instead of just doing a reflect.DeepEqual. An example where this is
 	// needed is *Set
-	if eq, ok := old.(Equal); ok {
-		return !eq.Equal(new)
+	if eq, ok := oldValue.(Equal); ok {
+		return !eq.Equal(newValue)
 	}
 
-	return !reflect.DeepEqual(old, new)
+	return !reflect.DeepEqual(oldValue, newValue)
 }
 
 // Id returns the ID of this resource.
@@ -506,16 +516,16 @@ func (d *ResourceDiff) GetRawPlan() cty.Value {
 // results from the exact levels for the new diff, then from state and diff as
 // per normal.
 func (d *ResourceDiff) getChange(key string) (getResult, getResult, bool) {
-	old := d.get(strings.Split(key, "."), "state")
-	var new getResult
+	oldValue := d.get(strings.Split(key, "."), "state")
+	var newValue getResult
 	for p := range d.updatedKeys {
 		if childAddrOf(key, p) {
-			new = d.getExact(strings.Split(key, "."), "newDiff")
-			return old, new, true
+			newValue = d.getExact(strings.Split(key, "."), "newDiff")
+			return oldValue, newValue, true
 		}
 	}
-	new = d.get(strings.Split(key, "."), "newDiff")
-	return old, new, false
+	newValue = d.get(strings.Split(key, "."), "newDiff")
+	return oldValue, newValue, false
 }
 
 // removed checks to see if the key is present in the existing, pre-customized
