@@ -1,6 +1,7 @@
 package schema
 
 import (
+	"context"
 	"fmt"
 	"reflect"
 	"sort"
@@ -38,7 +39,7 @@ type resourceDiffTestCase struct {
 }
 
 // testDiffCases produces a list of test cases for use with SetNew and SetDiff.
-func testDiffCases(t *testing.T, oldPrefix string, oldOffset int, computed bool) []resourceDiffTestCase {
+func testDiffCases(t *testing.T, computed bool) []resourceDiffTestCase {
 	return []resourceDiffTestCase{
 		{
 			Name: "basic primitive diff",
@@ -634,7 +635,7 @@ func testDiffCases(t *testing.T, oldPrefix string, oldOffset int, computed bool)
 }
 
 func TestSetNew(t *testing.T) {
-	testCases := testDiffCases(t, "", 0, false)
+	testCases := testDiffCases(t, false)
 	for _, tc := range testCases {
 		t.Run(tc.Name, func(t *testing.T) {
 			m := schemaMap(tc.Schema)
@@ -649,7 +650,7 @@ func TestSetNew(t *testing.T) {
 				return
 			}
 			for _, k := range d.UpdatedKeys() {
-				if err := m.diff(k, m[k], tc.Diff, d, false); err != nil {
+				if err := m.diff(context.Background(), k, m[k], tc.Diff, d, false); err != nil {
 					t.Fatalf("bad: %s", err)
 				}
 			}
@@ -661,7 +662,7 @@ func TestSetNew(t *testing.T) {
 }
 
 func TestSetNewComputed(t *testing.T) {
-	testCases := testDiffCases(t, "", 0, true)
+	testCases := testDiffCases(t, true)
 	for _, tc := range testCases {
 		t.Run(tc.Name, func(t *testing.T) {
 			m := schemaMap(tc.Schema)
@@ -676,7 +677,7 @@ func TestSetNewComputed(t *testing.T) {
 				return
 			}
 			for _, k := range d.UpdatedKeys() {
-				if err := m.diff(k, m[k], tc.Diff, d, false); err != nil {
+				if err := m.diff(context.Background(), k, m[k], tc.Diff, d, false); err != nil {
 					t.Fatalf("bad: %s", err)
 				}
 			}
@@ -945,7 +946,7 @@ func TestForceNew(t *testing.T) {
 				return
 			}
 			for _, k := range d.UpdatedKeys() {
-				if err := m.diff(k, m[k], tc.Diff, d, false); err != nil {
+				if err := m.diff(context.Background(), k, m[k], tc.Diff, d, false); err != nil {
 					t.Fatalf("bad: %s", err)
 				}
 			}
@@ -1194,7 +1195,7 @@ func TestClear(t *testing.T) {
 				return
 			}
 			for _, k := range d.UpdatedKeys() {
-				if err := m.diff(k, m[k], tc.Diff, d, false); err != nil {
+				if err := m.diff(context.Background(), k, m[k], tc.Diff, d, false); err != nil {
 					t.Fatalf("bad: %s", err)
 				}
 			}
@@ -1436,7 +1437,7 @@ func TestGetChangedKeysPrefix(t *testing.T) {
 			keys := d.GetChangedKeysPrefix(tc.Key)
 
 			for _, k := range d.UpdatedKeys() {
-				if err := m.diff(k, m[k], tc.Diff, d, false); err != nil {
+				if err := m.diff(context.Background(), k, m[k], tc.Diff, d, false); err != nil {
 					t.Fatalf("bad: %s", err)
 				}
 			}
@@ -1853,7 +1854,10 @@ func TestResourceDiffGetOkExistsSetNew(t *testing.T) {
 	}
 
 	d := newResourceDiff(tc.Schema, testConfig(t, map[string]interface{}{}), tc.State, tc.Diff)
-	d.SetNew(tc.Key, tc.Value)
+
+	if err := d.SetNew(tc.Key, tc.Value); err != nil {
+		t.Fatalf("unexpected SetNew error: %s", err)
+	}
 
 	v, ok := d.GetOkExists(tc.Key)
 	if s, ok := v.(*Set); ok {
@@ -1901,7 +1905,10 @@ func TestResourceDiffGetOkExistsSetNewComputed(t *testing.T) {
 	}
 
 	d := newResourceDiff(tc.Schema, testConfig(t, map[string]interface{}{}), tc.State, tc.Diff)
-	d.SetNewComputed(tc.Key)
+
+	if err := d.SetNewComputed(tc.Key); err != nil {
+		t.Fatalf("unexpected SetNewComputed error: %s", err)
+	}
 
 	_, ok := d.GetOkExists(tc.Key)
 
@@ -2142,7 +2149,10 @@ func TestResourceDiffNewValueKnownSetNew(t *testing.T) {
 	}
 
 	d := newResourceDiff(tc.Schema, tc.Config, tc.State, tc.Diff)
-	d.SetNew(tc.Key, tc.Value)
+
+	if err := d.SetNew(tc.Key, tc.Value); err != nil {
+		t.Fatalf("unexpected SetNew error: %s", err)
+	}
 
 	actual := d.NewValueKnown(tc.Key)
 	if tc.Expected != actual {
@@ -2179,10 +2189,117 @@ func TestResourceDiffNewValueKnownSetNewComputed(t *testing.T) {
 	}
 
 	d := newResourceDiff(tc.Schema, tc.Config, tc.State, tc.Diff)
-	d.SetNewComputed(tc.Key)
+
+	if err := d.SetNewComputed(tc.Key); err != nil {
+		t.Fatalf("unexpected SetNewComputed error: %s", err)
+	}
 
 	actual := d.NewValueKnown(tc.Key)
 	if tc.Expected != actual {
 		t.Fatalf("expected ok: %t, got: %t", tc.Expected, actual)
+	}
+}
+
+func TestResourceDiffHasChanges(t *testing.T) {
+	cases := []struct {
+		Schema map[string]*Schema
+		State  *terraform.InstanceState
+		Diff   *terraform.InstanceDiff
+		Keys   []string
+		Change bool
+	}{
+		// empty call d.HasChanges()
+		{
+			Schema: map[string]*Schema{},
+
+			State: nil,
+
+			Diff: &terraform.InstanceDiff{
+				Attributes: map[string]*terraform.ResourceAttrDiff{},
+			},
+
+			Keys: []string{},
+
+			Change: false,
+		},
+		// neither has change
+		{
+			Schema: map[string]*Schema{
+				"a": {
+					Type: TypeString,
+				},
+				"b": {
+					Type: TypeString,
+				},
+			},
+
+			State: &terraform.InstanceState{
+				Attributes: map[string]string{
+					"a": "foo",
+					"b": "foo",
+				},
+			},
+
+			Diff: &terraform.InstanceDiff{
+				Attributes: map[string]*terraform.ResourceAttrDiff{
+					"a": {
+						Old: "",
+						New: "foo",
+					},
+					"b": {
+						Old: "",
+						New: "foo",
+					},
+				},
+			},
+
+			Keys: []string{"a", "b"},
+
+			Change: false,
+		},
+		// one key has change
+		{
+			Schema: map[string]*Schema{
+				"a": {
+					Type: TypeString,
+				},
+				"b": {
+					Type: TypeString,
+				},
+			},
+
+			State: &terraform.InstanceState{
+				Attributes: map[string]string{
+					"a": "foo",
+					"b": "foo",
+				},
+			},
+
+			Diff: &terraform.InstanceDiff{
+				Attributes: map[string]*terraform.ResourceAttrDiff{
+					"a": {
+						Old: "",
+						New: "bar",
+					},
+					"b": {
+						Old: "",
+						New: "foo",
+					},
+				},
+			},
+
+			Keys: []string{"a", "b"},
+
+			Change: true,
+		},
+	}
+
+	for i, tc := range cases {
+		d := newResourceDiff(tc.Schema, testConfig(t, map[string]interface{}{}), tc.State, tc.Diff)
+
+		actual := d.HasChanges(tc.Keys...)
+		if actual != tc.Change {
+			t.Fatalf("Bad: %d %#v", i, actual)
+		}
 	}
 }
