@@ -6,13 +6,13 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func SuppressJsonDiff(k, old, new string, d *schema.ResourceData) bool {
-	oldMap, err := ExpandJsonFromString(old)
+func SuppressJsonDiff(k, oldValue, newValue string, d *schema.ResourceData) bool {
+	oldMap, err := ExpandJsonFromString(oldValue)
 	if err != nil {
 		return false
 	}
 
-	newMap, err := ExpandJsonFromString(new)
+	newMap, err := ExpandJsonFromString(newValue)
 	if err != nil {
 		return false
 	}
