@@ -18,12 +18,12 @@ func TestValidationIntBetween(t *testing.T) {
 		{
 			val:         1,
 			f:           IntBetween(2, 3),
-			expectedErr: regexp.MustCompile("expected [\\w]+ to be in the range \\(2 - 3\\), got 1"),
+			expectedErr: regexp.MustCompile(`expected [\w]+ to be in the range \(2 - 3\), got 1`),
 		},
 		{
 			val:         "1",
 			f:           IntBetween(2, 3),
-			expectedErr: regexp.MustCompile("expected type of [\\w]+ to be integer"),
+			expectedErr: regexp.MustCompile(`expected type of [\w]+ to be integer`),
 		},
 	})
 }
@@ -41,12 +41,12 @@ func TestValidationIntAtLeast(t *testing.T) {
 		{
 			val:         1,
 			f:           IntAtLeast(2),
-			expectedErr: regexp.MustCompile("expected [\\w]+ to be at least \\(2\\), got 1"),
+			expectedErr: regexp.MustCompile(`expected [\w]+ to be at least \(2\), got 1`),
 		},
 		{
 			val:         "1",
 			f:           IntAtLeast(2),
-			expectedErr: regexp.MustCompile("expected type of [\\w]+ to be integer"),
+			expectedErr: regexp.MustCompile(`expected type of [\w]+ to be integer`),
 		},
 	})
 }
@@ -64,12 +64,12 @@ func TestValidationIntAtMost(t *testing.T) {
 		{
 			val:         1,
 			f:           IntAtMost(0),
-			expectedErr: regexp.MustCompile("expected [\\w]+ to be at most \\(0\\), got 1"),
+			expectedErr: regexp.MustCompile(`expected [\w]+ to be at most \(0\), got 1`),
 		},
 		{
 			val:         "1",
 			f:           IntAtMost(0),
-			expectedErr: regexp.MustCompile("expected type of [\\w]+ to be integer"),
+			expectedErr: regexp.MustCompile(`expected type of [\w]+ to be integer`),
 		},
 	})
 }
@@ -119,12 +119,12 @@ func TestValidationIntInSlice(t *testing.T) {
 		{
 			val:         42,
 			f:           IntInSlice([]int{10, 20}),
-			expectedErr: regexp.MustCompile("expected [\\w]+ to be one of \\[10 20\\], got 42"),
+			expectedErr: regexp.MustCompile(`expected [\w]+ to be one of \[10 20\], got 42`),
 		},
 		{
 			val:         "InvalidValue",
 			f:           IntInSlice([]int{10, 20}),
-			expectedErr: regexp.MustCompile("expected type of [\\w]+ to be integer"),
+			expectedErr: regexp.MustCompile(`expected type of [\w]+ to be integer`),
 		},
 	})
 }
