@@ -1,8 +1,8 @@
 package convert
 
 import (
+	"context"
 	"fmt"
-	"log"
 	"reflect"
 	"sort"
 
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/configs/configschema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/logging"
 )
 
 func tftypeFromCtyType(in cty.Type) (tftypes.Type, error) {
@@ -128,10 +129,10 @@ func ctyTypeFromTFType(in tftypes.Type) (cty.Type, error) {
 
 // ConfigSchemaToProto takes a *configschema.Block and converts it to a
 // tfprotov5.SchemaBlock for a grpc response.
-func ConfigSchemaToProto(b *configschema.Block) *tfprotov5.SchemaBlock {
+func ConfigSchemaToProto(ctx context.Context, b *configschema.Block) *tfprotov5.SchemaBlock {
 	block := &tfprotov5.SchemaBlock{
 		Description:     b.Description,
-		DescriptionKind: protoStringKind(b.DescriptionKind),
+		DescriptionKind: protoStringKind(ctx, b.DescriptionKind),
 		Deprecated:      b.Deprecated,
 	}
 
@@ -141,7 +142,7 @@ func ConfigSchemaToProto(b *configschema.Block) *tfprotov5.SchemaBlock {
 		attr := &tfprotov5.SchemaAttribute{
 			Name:            name,
 			Description:     a.Description,
-			DescriptionKind: protoStringKind(a.DescriptionKind),
+			DescriptionKind: protoStringKind(ctx, a.DescriptionKind),
 			Optional:        a.Optional,
 			Computed:        a.Computed,
 			Required:        a.Required,
@@ -160,16 +161,16 @@ func ConfigSchemaToProto(b *configschema.Block) *tfprotov5.SchemaBlock {
 
 	for _, name := range sortedKeys(b.BlockTypes) {
 		b := b.BlockTypes[name]
-		block.BlockTypes = append(block.BlockTypes, protoSchemaNestedBlock(name, b))
+		block.BlockTypes = append(block.BlockTypes, protoSchemaNestedBlock(ctx, name, b))
 	}
 
 	return block
 }
 
-func protoStringKind(k configschema.StringKind) tfprotov5.StringKind {
+func protoStringKind(ctx context.Context, k configschema.StringKind) tfprotov5.StringKind {
 	switch k {
 	default:
-		log.Printf("[TRACE] unexpected configschema.StringKind: %d", k)
+		logging.HelperSchemaTrace(ctx, fmt.Sprintf("Unexpected configschema.StringKind: %d", k))
 		return tfprotov5.StringKindPlain
 	case configschema.StringPlain:
 		return tfprotov5.StringKindPlain
@@ -178,7 +179,7 @@ func protoStringKind(k configschema.StringKind) tfprotov5.StringKind {
 	}
 }
 
-func protoSchemaNestedBlock(name string, b *configschema.NestedBlock) *tfprotov5.SchemaNestedBlock {
+func protoSchemaNestedBlock(ctx context.Context, name string, b *configschema.NestedBlock) *tfprotov5.SchemaNestedBlock {
 	var nesting tfprotov5.SchemaNestedBlockNestingMode
 	switch b.Nesting {
 	case configschema.NestingSingle:
@@ -196,7 +197,7 @@ func protoSchemaNestedBlock(name string, b *configschema.NestedBlock) *tfprotov5
 	}
 	return &tfprotov5.SchemaNestedBlock{
 		TypeName: name,
-		Block:    ConfigSchemaToProto(&b.Block),
+		Block:    ConfigSchemaToProto(ctx, &b.Block),
 		Nesting:  nesting,
 		MinItems: int64(b.MinItems),
 		MaxItems: int64(b.MaxItems),
@@ -205,20 +206,20 @@ func protoSchemaNestedBlock(name string, b *configschema.NestedBlock) *tfprotov5
 
 // ProtoToConfigSchema takes the GetSchema_Block from a grpc response and converts it
 // to a terraform *configschema.Block.
-func ProtoToConfigSchema(b *tfprotov5.SchemaBlock) *configschema.Block {
+func ProtoToConfigSchema(ctx context.Context, b *tfprotov5.SchemaBlock) *configschema.Block {
 	block := &configschema.Block{
 		Attributes: make(map[string]*configschema.Attribute),
 		BlockTypes: make(map[string]*configschema.NestedBlock),
 
 		Description:     b.Description,
-		DescriptionKind: schemaStringKind(b.DescriptionKind),
+		DescriptionKind: schemaStringKind(ctx, b.DescriptionKind),
 		Deprecated:      b.Deprecated,
 	}
 
 	for _, a := range b.Attributes {
 		attr := &configschema.Attribute{
 			Description:     a.Description,
-			DescriptionKind: schemaStringKind(a.DescriptionKind),
+			DescriptionKind: schemaStringKind(ctx, a.DescriptionKind),
 			Required:        a.Required,
 			Optional:        a.Optional,
 			Computed:        a.Computed,
@@ -236,16 +237,16 @@ func ProtoToConfigSchema(b *tfprotov5.SchemaBlock) *configschema.Block {
 	}
 
 	for _, b := range b.BlockTypes {
-		block.BlockTypes[b.TypeName] = schemaNestedBlock(b)
+		block.BlockTypes[b.TypeName] = schemaNestedBlock(ctx, b)
 	}
 
 	return block
 }
 
-func schemaStringKind(k tfprotov5.StringKind) configschema.StringKind {
+func schemaStringKind(ctx context.Context, k tfprotov5.StringKind) configschema.StringKind {
 	switch k {
 	default:
-		log.Printf("[TRACE] unexpected tfprotov5.StringKind: %d", k)
+		logging.HelperSchemaTrace(ctx, fmt.Sprintf("Unexpected tfprotov5.StringKind: %d", k))
 		return configschema.StringPlain
 	case tfprotov5.StringKindPlain:
 		return configschema.StringPlain
@@ -254,7 +255,7 @@ func schemaStringKind(k tfprotov5.StringKind) configschema.StringKind {
 	}
 }
 
-func schemaNestedBlock(b *tfprotov5.SchemaNestedBlock) *configschema.NestedBlock {
+func schemaNestedBlock(ctx context.Context, b *tfprotov5.SchemaNestedBlock) *configschema.NestedBlock {
 	var nesting configschema.NestingMode
 	switch b.Nesting {
 	case tfprotov5.SchemaNestedBlockNestingModeSingle:
@@ -278,7 +279,7 @@ func schemaNestedBlock(b *tfprotov5.SchemaNestedBlock) *configschema.NestedBlock
 		MaxItems: int(b.MaxItems),
 	}
 
-	nested := ProtoToConfigSchema(b.Block)
+	nested := ProtoToConfigSchema(ctx, b.Block)
 	nb.Block = *nested
 	return nb
 }
