@@ -1,6 +1,7 @@
 package convert
 
 import (
+	"context"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -186,7 +187,7 @@ func TestConvertSchemaBlocks(t *testing.T) {
 
 	for name, tc := range tests {
 		t.Run(name, func(t *testing.T) {
-			converted := ProtoToConfigSchema(tc.Block)
+			converted := ProtoToConfigSchema(context.Background(), tc.Block)
 			if !cmp.Equal(converted, tc.Want, typeComparer, valueComparer, equateEmpty) {
 				t.Fatal(cmp.Diff(converted, tc.Want, typeComparer, valueComparer, equateEmpty))
 			}
@@ -362,7 +363,7 @@ func TestConvertProtoSchemaBlocks(t *testing.T) {
 
 	for name, tc := range tests {
 		t.Run(name, func(t *testing.T) {
-			converted := ConfigSchemaToProto(tc.Block)
+			converted := ConfigSchemaToProto(context.Background(), tc.Block)
 			if !cmp.Equal(converted, tc.Want, typeComparer, equateEmpty) {
 				t.Fatal(cmp.Diff(converted, tc.Want, typeComparer, equateEmpty))
 			}
