@@ -14,6 +14,7 @@ import (
 	"github.com/hashicorp/hc-install/product"
 	"github.com/hashicorp/hc-install/releases"
 	"github.com/hashicorp/hc-install/src"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/logging"
 )
 
 // Config is used to configure the test helper. In most normal test programs
@@ -29,11 +30,11 @@ type Config struct {
 
 // DiscoverConfig uses environment variables and other means to automatically
 // discover a reasonable test helper configuration.
-func DiscoverConfig(sourceDir string) (*Config, error) {
-	tfVersion := strings.TrimPrefix(os.Getenv("TF_ACC_TERRAFORM_VERSION"), "v")
-	tfPath := os.Getenv("TF_ACC_TERRAFORM_PATH")
+func DiscoverConfig(ctx context.Context, sourceDir string) (*Config, error) {
+	tfVersion := strings.TrimPrefix(os.Getenv(EnvTfAccTerraformVersion), "v")
+	tfPath := os.Getenv(EnvTfAccTerraformPath)
 
-	tempDir := os.Getenv("TF_ACC_TEMP_DIR")
+	tempDir := os.Getenv(EnvTfAccTempDir)
 	tfDir, err := ioutil.TempDir(tempDir, "plugintest-terraform")
 	if err != nil {
 		return nil, fmt.Errorf("failed to create temp dir: %w", err)
@@ -42,6 +43,8 @@ func DiscoverConfig(sourceDir string) (*Config, error) {
 	var sources []src.Source
 	switch {
 	case tfPath != "":
+		logging.HelperResourceTrace(ctx, fmt.Sprintf("Adding potential Terraform CLI source of exact path: %s", tfPath))
+
 		sources = append(sources, &fs.AnyVersion{
 			ExactBinPath: tfPath,
 		})
@@ -52,12 +55,17 @@ func DiscoverConfig(sourceDir string) (*Config, error) {
 			return nil, fmt.Errorf("invalid Terraform version: %w", err)
 		}
 
+		logging.HelperResourceTrace(ctx, fmt.Sprintf("Adding potential Terraform CLI source of releases.hashicorp.com exact version %q for installation in: %s", tfVersion, tfDir))
+
 		sources = append(sources, &releases.ExactVersion{
 			InstallDir: tfDir,
 			Product:    product.Terraform,
 			Version:    tfVersion,
 		})
 	default:
+		logging.HelperResourceTrace(ctx, "Adding potential Terraform CLI source of local filesystem PATH lookup")
+		logging.HelperResourceTrace(ctx, fmt.Sprintf("Adding potential Terraform CLI source of checkpoint.hashicorp.com latest version for installation in: %s", tfDir))
+
 		sources = append(sources, &fs.AnyVersion{
 			Product: &product.Terraform,
 		})
@@ -70,8 +78,12 @@ func DiscoverConfig(sourceDir string) (*Config, error) {
 	installer := install.NewInstaller()
 	tfExec, err := installer.Ensure(context.Background(), sources)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to find or install Terraform CLI from %+v: %w", sources, err)
 	}
+
+	ctx = logging.TestTerraformPathContext(ctx, tfExec)
+
+	logging.HelperResourceDebug(ctx, "Found Terraform CLI")
 
 	return &Config{
 		SourceDir:     sourceDir,
