@@ -1,15 +1,16 @@
 package plugintest
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
+	"strings"
 
 	"github.com/hashicorp/terraform-exec/tfexec"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/logging"
 )
-
-const subprocessCurrentSigil = "4acd63807899403ca4859f5bb948d2c6"
-const subprocessPreviousSigil = "2279afb8cf71423996be1fd65d32f13b"
 
 // AutoInitProviderHelper is the main entrypoint for testing provider plugins
 // using this package. It is intended to be called during TestMain to prepare
@@ -20,8 +21,8 @@ const subprocessPreviousSigil = "2279afb8cf71423996be1fd65d32f13b"
 // available for upgrade tests, and then will return an object containing the
 // results of that initialization which can then be stored in a global variable
 // for use in other tests.
-func AutoInitProviderHelper(sourceDir string) *Helper {
-	helper, err := AutoInitHelper(sourceDir)
+func AutoInitProviderHelper(ctx context.Context, sourceDir string) *Helper {
+	helper, err := AutoInitHelper(ctx, sourceDir)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "cannot run Terraform provider tests: %s\n", err)
 		os.Exit(1)
@@ -49,13 +50,13 @@ type Helper struct {
 // way to get the standard init behavior based on environment variables, and
 // callers should use this unless they have an unusual requirement that calls
 // for constructing a config in a different way.
-func AutoInitHelper(sourceDir string) (*Helper, error) {
-	config, err := DiscoverConfig(sourceDir)
+func AutoInitHelper(ctx context.Context, sourceDir string) (*Helper, error) {
+	config, err := DiscoverConfig(ctx, sourceDir)
 	if err != nil {
 		return nil, err
 	}
 
-	return InitHelper(config)
+	return InitHelper(ctx, config)
 }
 
 // InitHelper prepares a testing helper with the given configuration.
@@ -67,8 +68,8 @@ func AutoInitHelper(sourceDir string) (*Helper, error) {
 // If this function returns an error then it may have left some temporary files
 // behind in the system's temporary directory. There is currently no way to
 // automatically clean those up.
-func InitHelper(config *Config) (*Helper, error) {
-	tempDir := os.Getenv("TF_ACC_TEMP_DIR")
+func InitHelper(ctx context.Context, config *Config) (*Helper, error) {
+	tempDir := os.Getenv(EnvTfAccTempDir)
 	baseDir, err := ioutil.TempDir(tempDir, "plugintest")
 	if err != nil {
 		return nil, fmt.Errorf("failed to create temporary directory for test helper: %s", err)
@@ -103,22 +104,153 @@ func (h *Helper) Close() error {
 // If the working directory object is not itself closed by the time the test
 // program exits, the Close method on the helper itself will attempt to
 // delete it.
-func (h *Helper) NewWorkingDir() (*WorkingDir, error) {
+func (h *Helper) NewWorkingDir(ctx context.Context, t TestControl) (*WorkingDir, error) {
 	dir, err := ioutil.TempDir(h.baseDir, "work")
 	if err != nil {
 		return nil, err
 	}
 
+	ctx = logging.TestWorkingDirectoryContext(ctx, dir)
+
 	// symlink the provider source files into the config directory
 	// e.g. testdata
+	logging.HelperResourceTrace(ctx, "Symlinking source directories to work directory")
 	err = symlinkDirectoriesOnly(h.sourceDir, dir)
 	if err != nil {
 		return nil, err
 	}
 
 	tf, err := tfexec.NewTerraform(dir, h.terraformExec)
+
 	if err != nil {
+		return nil, fmt.Errorf("unable to create terraform-exec instance: %w", err)
+	}
+
+	err = tf.SetDisablePluginTLS(true)
+
+	if err != nil {
+		return nil, fmt.Errorf("unable to disable terraform-exec plugin TLS: %w", err)
+	}
+
+	err = tf.SetSkipProviderVerify(true) // Only required for Terraform CLI 0.12.x
+
+	var mismatch *tfexec.ErrVersionMismatch
+	if err != nil && !errors.As(err, &mismatch) {
+		return nil, fmt.Errorf("unable to disable terraform-exec provider verification: %w", err)
+	}
+
+	tfAccLog := os.Getenv(EnvTfAccLog)
+	tfAccLogPath := os.Getenv(EnvTfAccLogPath)
+	tfLogCore := os.Getenv(EnvTfLogCore)
+	tfLogPathMask := os.Getenv(EnvTfLogPathMask)
+	tfLogProvider := os.Getenv(EnvTfLogProvider)
+
+	if tfAccLog != "" && tfLogCore != "" {
+		err = fmt.Errorf(
+			"Invalid environment variable configuration. Cannot set both TF_ACC_LOG and TF_LOG_CORE. " +
+				"Use TF_LOG_CORE and TF_LOG_PROVIDER to separately control the Terraform CLI logging subsystems. " +
+				"To control the Go standard library log package for the provider under test, use TF_LOG.",
+		)
+		logging.HelperResourceError(ctx, err.Error())
 		return nil, err
+	}
+
+	if tfAccLog != "" {
+		logging.HelperResourceTrace(
+			ctx,
+			fmt.Sprintf("Setting terraform-exec log level via %s environment variable, if Terraform CLI is version 0.15 or later", EnvTfAccLog),
+			map[string]interface{}{logging.KeyTestTerraformLogLevel: tfAccLog},
+		)
+
+		err := tf.SetLog(tfAccLog)
+
+		if err != nil {
+			if !errors.As(err, new(*tfexec.ErrVersionMismatch)) {
+				logging.HelperResourceError(
+					ctx,
+					"Unable to set terraform-exec log level",
+					map[string]interface{}{logging.KeyError: err.Error()},
+				)
+				return nil, fmt.Errorf("unable to set terraform-exec log level (%s): %w", tfAccLog, err)
+			}
+
+			logging.HelperResourceWarn(
+				ctx,
+				fmt.Sprintf("Unable to set terraform-exec log level via %s environment variable, as Terraform CLI is version 0.14 or earlier. It will default to TRACE.", EnvTfAccLog),
+				map[string]interface{}{logging.KeyTestTerraformLogLevel: "TRACE"},
+			)
+		}
+	}
+
+	if tfLogCore != "" {
+		logging.HelperResourceTrace(
+			ctx,
+			fmt.Sprintf("Setting terraform-exec core log level via %s environment variable, if Terraform CLI is version 0.15 or later", EnvTfLogCore),
+			map[string]interface{}{
+				logging.KeyTestTerraformLogCoreLevel: tfLogCore,
+			},
+		)
+
+		err := tf.SetLogCore(tfLogCore)
+
+		if err != nil {
+			logging.HelperResourceError(
+				ctx,
+				"Unable to set terraform-exec core log level",
+				map[string]interface{}{logging.KeyError: err.Error()},
+			)
+			return nil, fmt.Errorf("unable to set terraform-exec core log level (%s): %w", tfLogCore, err)
+		}
+	}
+
+	if tfLogProvider != "" {
+		logging.HelperResourceTrace(
+			ctx,
+			fmt.Sprintf("Setting terraform-exec provider log level via %s environment variable, if Terraform CLI is version 0.15 or later", EnvTfLogProvider),
+			map[string]interface{}{
+				logging.KeyTestTerraformLogCoreLevel: tfLogProvider,
+			},
+		)
+
+		err := tf.SetLogProvider(tfLogProvider)
+
+		if err != nil {
+			logging.HelperResourceError(
+				ctx,
+				"Unable to set terraform-exec provider log level",
+				map[string]interface{}{logging.KeyError: err.Error()},
+			)
+			return nil, fmt.Errorf("unable to set terraform-exec provider log level (%s): %w", tfLogProvider, err)
+		}
+	}
+
+	var logPath, logPathEnvVar string
+
+	if tfAccLogPath != "" {
+		logPath = tfAccLogPath
+		logPathEnvVar = EnvTfAccLogPath
+	}
+
+	// Similar to helper/logging.LogOutput() and
+	// terraform-plugin-log/tfsdklog.RegisterTestSink(), the TF_LOG_PATH_MASK
+	// environment variable should take precedence over TF_ACC_LOG_PATH.
+	if tfLogPathMask != "" {
+		// Escape special characters which may appear if we have subtests
+		testName := strings.Replace(t.Name(), "/", "__", -1)
+		logPath = fmt.Sprintf(tfLogPathMask, testName)
+		logPathEnvVar = EnvTfLogPathMask
+	}
+
+	if logPath != "" {
+		logging.HelperResourceTrace(
+			ctx,
+			fmt.Sprintf("Setting terraform-exec log path via %s environment variable", logPathEnvVar),
+			map[string]interface{}{logging.KeyTestTerraformLogPath: logPath},
+		)
+
+		if err := tf.SetLogPath(logPath); err != nil {
+			return nil, fmt.Errorf("unable to set terraform-exec log path (%s): %w", logPath, err)
+		}
 	}
 
 	return &WorkingDir{
@@ -132,16 +264,21 @@ func (h *Helper) NewWorkingDir() (*WorkingDir, error) {
 // RequireNewWorkingDir is a variant of NewWorkingDir that takes a TestControl
 // object and will immediately fail the running test if the creation of the
 // working directory fails.
-func (h *Helper) RequireNewWorkingDir(t TestControl) *WorkingDir {
+func (h *Helper) RequireNewWorkingDir(ctx context.Context, t TestControl) *WorkingDir {
 	t.Helper()
 
-	wd, err := h.NewWorkingDir()
+	wd, err := h.NewWorkingDir(ctx, t)
 	if err != nil {
 		t := testingT{t}
 		t.Fatalf("failed to create new working directory: %s", err)
 		return nil
 	}
 	return wd
+}
+
+// WorkingDirectory returns the working directory being used when running tests.
+func (h *Helper) WorkingDirectory() string {
+	return h.baseDir
 }
 
 // TerraformExecPath returns the location of the Terraform CLI executable that
